updated gameframwork
Moved mainActivity to gameHumanPlayer
