+++ app/src/main/java/com/example/game/Checkers/info/CheckerState.java
@@ -1,5 +1,8 @@
 package com.example.game.Checkers.info;
 
+import com.example.game.Checkers.BlackChecker;
+import com.example.game.Checkers.Checker;
+import com.example.game.Checkers.RedChecker;
 import com.example.game.GameFramework.info.GameState;
 import java.io.Serializable;
 
@@ -15,6 +18,8 @@ public class CheckerState extends GameState implements Serializable {
 
 
     private char[][] board;
+    public Checker[][] checkerList;
+
 
     private int playerToMove;
 
@@ -27,10 +32,21 @@ public class CheckerState extends GameState implements Serializable {
             }
         }
         playerToMove = 0;
+        checkerList = new Checker[][]
+                {{null, new RedChecker(0,1), null, new RedChecker(0,3), null, new RedChecker(0,5), null, new RedChecker(0,7)},
+                        {new RedChecker(1,0), null, new RedChecker(1,2),null, new RedChecker(1,4), null, new RedChecker(1,6), null},
+                        {null, new RedChecker(2,1), null, new RedChecker(2,3), null, new RedChecker(2,5), null, new RedChecker(2,7)},
+                        {null, null, null, null,null, null,null, null},
+                        {null, null, null, null,null, null,null, null},
+                        {new BlackChecker(5,0), null, new BlackChecker(5,2),null, new BlackChecker(5,4), null, new BlackChecker(5,6), null},
+                        {null, new BlackChecker(6,1), null, new BlackChecker(6,3), null, new BlackChecker(6,5), null, new BlackChecker(6,7)},
+                        {new BlackChecker(7,0), null, new BlackChecker(7,2),null, new BlackChecker(7,4), null, new BlackChecker(7,6), null}
+                };
     }
 
     public CheckerState(CheckerState orig) {
         super();
+        //TODO copy checkerList after creating new empty CheckerList
         board = new char[8][8];
         for (int i = 0; i < 8; i++){
             for (int j = 0; j < 8; j++){
@@ -59,6 +75,10 @@ public class CheckerState extends GameState implements Serializable {
 
         // return the character that is in the proper position
         board[row][col] = piece;
+    }
+
+    public Checker[][] getCheckerList() {
+        return checkerList;
     }
 
     //Tells whos move it is
